44. Delete confirmation dialog in Blazor

+++ EmployeeManagement.Web/Pages/DisplayEmployeeBase.cs
@@ -23,15 +23,26 @@ namespace EmployeeManagement.Web.Pages
         [Parameter]
         public EventCallback<int> OnEmployeeDeleted { get; set; }
 
+        protected Pragim.Components.ConfirmBase DeleteConfirmation { get; set; }
+
+
         [Inject]
         public IEmployeeService EmployeeService { get; set; }
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
+            }
+        }
 
         protected async Task Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
+            DeleteConfirmation.Show();
+            //await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+            //await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
             //NavigationManager.NavigateTo("/", true);
-
         }
 
         [Inject]
